Revert to using script from pypa/python-manylinux-demo
And use data_files instead of package_data in order to install to a .data
directory inside the wheel instead of purelib, which was wrong and was
making `auditwheel` fail.

Also removed the restrained_ESP_fit directory, as there's no need for
it. The wrapper script is now in repo root.