Test installing from sdist (again)
I still expect this to fail because the Fortran source is currently not
included in the sdist. This should be fixed when we start using
setuptools-scm (in order to derive the package version from git tags),
which also has the effect of including all files in version control in
the sdist (AFAIUI).